rewrite Input implementation

+++ gdx/src/com/badlogic/gdx/InputAdapter.java
@@ -54,7 +54,7 @@ public class InputAdapter implements InputProcessor {
 	}
 
 	@Override
-	public boolean scrolled (float amountX, float amountY) {
+	public boolean scrolled (double amountX, double amountY) {
 		return false;
 	}
 }
